+++ packages/app-data/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [4.6.12](https://github.com/cowprotocol/cow-sdk/compare/sdk-app-data-v4.6.11...sdk-app-data-v4.6.12) (2026-03-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.0
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.0
+
+## [4.6.11](https://github.com/cowprotocol/cow-sdk/compare/sdk-app-data-v4.6.10...sdk-app-data-v4.6.11) (2026-03-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.8.2
+  * devDependencies
+    * @cowprotocol/sdk-config bumped to 1.1.2
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.3.14
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.3.14
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.14
+
 ## [4.6.10](https://github.com/cowprotocol/cow-sdk/compare/sdk-app-data-v4.6.9...sdk-app-data-v4.6.10) (2026-03-16)