+++ packages/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.2](https://github.com/cowprotocol/cow-sdk/compare/sdk-config-v1.1.1...sdk-config-v1.1.2) (2026-03-17)
+
+
+### 🐛 Bug Fixes
+
+* use pnpm trusted publishing ([#841](https://github.com/cowprotocol/cow-sdk/issues/841)) ([1885b6f](https://github.com/cowprotocol/cow-sdk/commit/1885b6fa2a006670e3bb51a2ac8b60ed9c157eba))
+
 ## [1.1.1](https://github.com/cowprotocol/cow-sdk/compare/sdk-config-v1.1.0...sdk-config-v1.1.1) (2026-03-16)