+++ packages/cow-shed/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## [0.3.3](https://github.com/cowprotocol/cow-sdk/compare/sdk-cow-shed-v0.3.2...sdk-cow-shed-v0.3.3) (2026-03-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-contracts-ts bumped to 2.2.0
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.0
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.0
+
+## [0.3.2](https://github.com/cowprotocol/cow-sdk/compare/sdk-cow-shed-v0.3.1...sdk-cow-shed-v0.3.2) (2026-03-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-config bumped to 1.1.2
+    * @cowprotocol/sdk-common bumped to 0.8.2
+    * @cowprotocol/sdk-contracts-ts bumped to 2.1.2
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.3.14
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.3.14
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.14
+
 ## [0.3.1](https://github.com/cowprotocol/cow-sdk/compare/sdk-cow-shed-v0.3.0...sdk-cow-shed-v0.3.1) (2026-03-16)