+++ packages/providers/ethers-v6-adapter/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.4.0](https://github.com/cowprotocol/cow-sdk/compare/sdk-ethers-v6-adapter-v0.3.14...sdk-ethers-v6-adapter-v0.4.0) (2026-03-17)
+
+
+### ✨ Features
+
+* use address utils to compare addresses ([#827](https://github.com/cowprotocol/cow-sdk/issues/827)) ([50a66ff](https://github.com/cowprotocol/cow-sdk/commit/50a66ff044f98ce0313c6213e31c83c9028836a1))
+
+## [0.3.14](https://github.com/cowprotocol/cow-sdk/compare/sdk-ethers-v6-adapter-v0.3.13...sdk-ethers-v6-adapter-v0.3.14) (2026-03-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.8.2
+
 ## [0.3.13](https://github.com/cowprotocol/cow-sdk/compare/sdk-ethers-v6-adapter-v0.3.12...sdk-ethers-v6-adapter-v0.3.13) (2026-03-16)